requirements.txt
numpy
Pillow
opencv-python
tensorflow
matplotlib